+++ packages/api/lib/request.js
@@ -1,35 +1,11 @@
 const log = require('@cumulus/common/log');
-const { JsonWebTokenError, TokenExpiredError } = require('jsonwebtoken');
 
 const { AccessToken, User } = require('../models');
 const {
   TokenUnauthorizedUserError,
   TokenNotFoundError
 } = require('../lib/errors');
-const {
-  AuthorizationFailureResponse,
-  InvalidTokenResponse,
-  TokenExpiredResponse
-} = require('../lib/responses');
 const { verifyJwtToken } = require('./token');
-
-const handleRequestAuthorizationError = (err) => {
-  if (err instanceof TokenExpiredError) {
-    return new TokenExpiredResponse();
-  }
-  if (err instanceof JsonWebTokenError) {
-    return new InvalidTokenResponse();
-  }
-  if (err instanceof TokenUnauthorizedUserError) {
-    return new AuthorizationFailureResponse({
-      message: 'User not authorized',
-      statusCode: 403
-    });
-  }
-  if (err instanceof TokenNotFoundError) {
-    return new InvalidTokenResponse();
-  }
-}
 
 const verifyRequestAuthorization = async (requestJwtToken) => {
   let accessToken;
@@ -68,6 +44,5 @@ const verifyRequestAuthorization = async (requestJwtToken) => {
 };
 
 module.exports = {
-  handleRequestAuthorizationError,
   verifyRequestAuthorization
 };
